move TFPipeline into tfos dir
